+++ src/rx-query.js
@@ -15,11 +15,12 @@ import {
 } from 'rxjs';
 import {
     mergeMap,
-    filter
+    filter,
+    map
 } from 'rxjs/operators';
 
 let _queryCount = 0;
-const newQueryID = function() {
+const newQueryID = function () {
     return ++_queryCount;
 };
 
@@ -204,7 +205,7 @@ export class RxQuery {
                         else return results;
                     }),
                     filter(results => results !== 'WAITFORNEXTEMIT'),
-                )
+            )
                 .asObservable();
 
             // we also subscribe to the changeEvent-stream so it detects changed if it has subscribers
@@ -223,7 +224,14 @@ export class RxQuery {
                     changeEvents$
                 );
         }
-        return this._$;
+        return this._$
+            .pipe(
+                map(current => {
+                    // copy the array so it wont matter if the user modifies it
+                    const ret = Array.isArray(current) ? current.slice() : current;
+                    return ret;
+                })
+            );
     }
 
     toJSON() {
@@ -360,7 +368,11 @@ export class RxQuery {
             changed = await this._ensureEqual();
 
         // than return the current results
-        return this._results$.getValue();
+        const current = this._results$.getValue();
+
+        // copy the array so it wont matter if the user modifies it
+        const ret = Array.isArray(current) ? current.slice() : current;
+        return ret;
     }
 
     /**
@@ -453,12 +465,12 @@ export class RxQuery {
  * @param  {string[]} mQueryProtoKeys [description]
  * @return {void}                 [description]
  */
-const protoMerge = function(rxQueryProto, mQueryProtoKeys) {
+const protoMerge = function (rxQueryProto, mQueryProtoKeys) {
     mQueryProtoKeys
         .filter(attrName => !attrName.startsWith('_'))
         .filter(attrName => !rxQueryProto[attrName])
         .forEach(attrName => {
-            rxQueryProto[attrName] = function(p1) {
+            rxQueryProto[attrName] = function (p1) {
                 const clonedThis = this._clone();
                 clonedThis.mquery[attrName](p1);
                 return clonedThis._tunnelQueryCache();